+++ src/owlplanner/socialsecurity.py
@@ -18,8 +18,10 @@ import numpy as np
 
 def getFRAs(yobs):
     """
-    Return full retirement age based on birth year.
-    Returns an array of fractional age.
+    Return full retirement age based on birth year and day of the month.
+    Returns an array of fractional age in 1/12 increments, one for each individual.
+    Ages returned are in SS age, i.e., not in conventional age. Comparisons to FRA
+    will need to be done in SS age.
     """
     fras = np.zeros(len(yobs))
 
@@ -36,6 +38,7 @@ def getFRAs(yobs):
 def getSpousalBenefits(pias):
     """
     Compute spousal benefit. Returns an array.
+    Not affected by which day of the month is the birthday.
     """
     icount = len(pias)
     benefits = np.zeros(icount)
@@ -51,15 +54,20 @@ def getSpousalBenefits(pias):
     return benefits
 
 
-def getSelfFactor(fra, age):
+def getSelfFactor(fra, convage, bornOnFirst):
     """
     Return factor to multiply PIA given the age when SS starts.
-    Year of FRA and age can be fractional.
+    Year of FRA and conventional age can be fractional with 1/12 increments.
+    Assume that age has been corrected if birthday is on the first day of the month.
     """
-    if age < 62 or age > 70:
-        raise ValueError(f"Age {age} out of range.")
+    if convage < 62 or convage > 70:
+        raise ValueError(f"Age {convage} out of range.")
 
-    diff = fra - age
+    # Add a month to conventional age if born on the first.
+    offset = 0 if not bornOnFirst else 1/12
+    ssage = convage + offset
+
+    diff = fra - ssage
     if diff <= 0:
         return 1. - .08 * diff
     elif diff <= 3:
@@ -70,15 +78,19 @@ def getSelfFactor(fra, age):
         return .8 - 0.05 * (diff - 3)
 
 
-def getSpousalFactor(fra, age):
+def getSpousalFactor(fra, convage, bornOnFirst):
     """
-    Return factor to multiply spousal benefit given the age when benefit starts.
-    Year of FRA and age can be fractional.
+    Return factor to multiply spousal benefits given the age when benefits start.
+    Year of FRA and conventional age can be fractional with 1/12 increments.
     """
-    if age < 62:
-        raise ValueError(f"Age {age} out of range.")
+    if convage < 62:
+        raise ValueError(f"Age {convage} out of range.")
 
-    diff = fra - age
+    # Add a month to conventional age if born on the first.
+    offset = 0 if not bornOnFirst else 1/12
+    ssage = convage + offset
+
+    diff = fra - ssage
     if diff <= 0:
         return 1.
     elif diff <= 3: