+++ tests/test_repro.py
@@ -14,10 +14,9 @@ thisyear = date.today().year
 
 def createJackAndJillPlan(name):
     inames = ['Jack', 'Jill']
-    yobs = [1963, 1966]
-    mobs = [1, 1]
+    dobs = ["1963-01-15", "1966-01-16"]
     expectancy = [thisyear - 1963 + 20, thisyear - 1966 + 20]
-    p = owl.Plan(inames, yobs, mobs, expectancy, name)
+    p = owl.Plan(inames, dobs, expectancy, name)
     p.setSpendingProfile('flat', 60)
 
     p.setAccountBalances(taxable=[90, 60], taxDeferred=[600, 150],