+++ tests/test_socsec.py
@@ -25,16 +25,19 @@ def test_selfFactor():
     factors66 = [0.75, 0.80, 0.866667, 0.9333333, 1.0, 1.08, 1.16, 1.24, 1.32]
     factors67 = [0.70, 0.75, 0.80, 0.866667, 0.9333333, 1.0, 1.08, 1.16, 1.24]
     for i, a in enumerate(ages):
-        assert ss.getSelfFactor(66, a) == pytest.approx(factors66[i], 0.001)
-        assert ss.getSelfFactor(67, a) == pytest.approx(factors67[i], 0.001)
+        assert ss.getSelfFactor(66, a, False) == pytest.approx(factors66[i], 0.001)
+        assert ss.getSelfFactor(67, a, False) == pytest.approx(factors67[i], 0.001)
+        if a > 62:
+            assert ss.getSelfFactor(66, a - 1/12, True) == pytest.approx(factors66[i], 0.001)
+            assert ss.getSelfFactor(67, a - 1/12, True) == pytest.approx(factors67[i], 0.001)
 
     # Example from SSA: https://www.ssa.gov/benefits/retirement/planner/1955-delay.html
-    assert ss.getSelfFactor(66 + 2/12, 66 + 2/12) == pytest.approx(1.00, 0.001)
-    assert ss.getSelfFactor(66 + 2/12, 67) == pytest.approx(1.06667, 0.001)
-    assert ss.getSelfFactor(66 + 2/12, 68) == pytest.approx(1.14667, 0.001)
-    assert ss.getSelfFactor(66 + 2/12, 69) == pytest.approx(1.22667, 0.001)
-    assert ss.getSelfFactor(66 + 3/12, 69 + 1/12) == pytest.approx(1.22667, 0.001)
-    assert ss.getSelfFactor(66 + 2/12, 70) == pytest.approx(1.30667, 0.001)
+    assert ss.getSelfFactor(66 + 2/12, 66 + 2/12, False) == pytest.approx(1.00, 0.001)
+    assert ss.getSelfFactor(66 + 2/12, 67, False) == pytest.approx(1.06667, 0.001)
+    assert ss.getSelfFactor(66 + 2/12, 68, False) == pytest.approx(1.14667, 0.001)
+    assert ss.getSelfFactor(66 + 2/12, 69, False) == pytest.approx(1.22667, 0.001)
+    assert ss.getSelfFactor(66 + 3/12, 69 + 1/12, False) == pytest.approx(1.22667, 0.001)
+    assert ss.getSelfFactor(66 + 2/12, 70, False) == pytest.approx(1.30667, 0.001)
 
 
 def test_spousalFactor():
@@ -42,16 +45,19 @@ def test_spousalFactor():
     factors66 = [0.70, 0.75, 0.833333, 0.9166667, 1.0, 1.0, 1.0, 1.0, 1.0]
     factors67 = [0.65, 0.70, 0.75, 0.833333, 0.9166667, 1.0, 1.0, 1.0, 1.0]
     for i, a in enumerate(ages):
-        assert ss.getSpousalFactor(66, a) == pytest.approx(factors66[i], 0.001)
-        assert ss.getSpousalFactor(67, a) == pytest.approx(factors67[i], 0.001)
+        assert ss.getSpousalFactor(66, a, False) == pytest.approx(factors66[i], 0.001)
+        assert ss.getSpousalFactor(67, a, False) == pytest.approx(factors67[i], 0.001)
+        if a > 62:
+            assert ss.getSpousalFactor(66, a - 1/12, True) == pytest.approx(factors66[i], 0.001)
+            assert ss.getSpousalFactor(67, a - 1/12, True) == pytest.approx(factors67[i], 0.001)
 
     # Individual born in 1955.
-    assert ss.getSpousalFactor(66 + 2/12, 66 + 2/12) == pytest.approx(1.00, 0.001)
-    assert ss.getSpousalFactor(66 + 2/12, 66) == pytest.approx(2*0.4931, 0.001)
-    assert ss.getSpousalFactor(66 + 2/12, 65) == pytest.approx(2*0.4514, 0.001)
-    assert ss.getSpousalFactor(66 + 2/12, 64) == pytest.approx(2*0.4097, 0.001)
-    assert ss.getSpousalFactor(66 + 2/12, 63) == pytest.approx(2*0.3708, 0.001)
-    assert ss.getSpousalFactor(66 + 2/12, 62) == pytest.approx(2*0.3458, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 66 + 2/12, False) == pytest.approx(1.00, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 66, False) == pytest.approx(2*0.4931, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 65, False) == pytest.approx(2*0.4514, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 64, False) == pytest.approx(2*0.4097, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 63, False) == pytest.approx(2*0.3708, 0.001)
+    assert ss.getSpousalFactor(66 + 2/12, 62, False) == pytest.approx(2*0.3458, 0.001)
 
 
 def test_SpousalBenefits():