+++ ui/Create_Case.py
@@ -90,15 +90,12 @@ else:
         if iname0:
             incol1, incol2 = st.columns(2, gap="large", vertical_alignment="top")
             with incol1:
-                kz.initCaseKey("yob0", 1965)
-                ret = kz.getIntNum(f"{iname0}'s birth year", "yob0", disabled=diz2)
+                kz.initCaseKey("dob0", "1965-01-15")
+                ret = kz.getDate(f"{iname0}'s date of birth", "dob0", disabled=diz2)
 
             with incol2:
-                kz.initCaseKey("mob0", 1)
-                ret = kz.getIntNum(f"{iname0}'s birth month", "mob0", min_value=1, max_value=12, disabled=diz2)
-
-            kz.initCaseKey("life0", 80)
-            ret = kz.getIntNum(f"{iname0}'s expected longevity", "life0", disabled=diz1)
+                kz.initCaseKey("life0", 80)
+                ret = kz.getIntNum(f"{iname0}'s expected longevity", "life0", disabled=diz1)
 
     with col2:
         if kz.getCaseKey("status") == "married":
@@ -111,15 +108,12 @@ else:
             if iname1:
                 incol1, incol2 = st.columns(2, gap="large", vertical_alignment="top")
                 with incol1:
-                    kz.initCaseKey("yob1", 1965)
-                    ret = kz.getIntNum(f"{iname1}'s birth year", "yob1", disabled=diz2)
+                    kz.initCaseKey("dob1", "1965-01-15")
+                    ret = kz.getDate(f"{iname1}'s date of birth", "dob1", disabled=diz2)
 
                 with incol2:
-                    kz.initCaseKey("mob1", 1)
-                    ret = kz.getIntNum(f"{iname1}'s birth month", "mob1", min_value=1, max_value=12, disabled=diz2)
-
-                kz.initCaseKey("life1", 80)
-                ret = kz.getIntNum(f"{iname1}'s expected longevity", "life1", disabled=diz1)
+                    kz.initCaseKey("life1", 80)
+                    ret = kz.getIntNum(f"{iname1}'s expected longevity", "life1", disabled=diz1)
 
     st.divider()
     cantcreate = kz.isIncomplete() or diz1