+++ ui/Fixed_Income.py
@@ -3,7 +3,7 @@ import streamlit as st
 import sskeys as kz
 
 
-def getIntInput(i, key, thing, defval=0, helpmsg=None, max_val=None, prompt=True):
+def getIntInput(i, key, thing, defval=0, helpmsg=None, min_val=0, max_val=None, prompt=True):
     nkey = key + str(i)
     kz.initCaseKey(nkey, defval)
     if prompt:
@@ -11,7 +11,7 @@ def getIntInput(i, key, thing, defval=0, helpmsg=None, max_val=None, prompt=True
     else:
         own = ""
     return st.number_input(
-        f"{own}{thing}", min_value=0, value=kz.getCaseKey(nkey),
+        f"{own}{thing}", min_value=min_val, value=kz.getCaseKey(nkey),
         on_change=kz.setpull, help=helpmsg, args=[nkey], key=kz.genCaseKey(nkey),
         max_value=max_val,
     )
@@ -49,31 +49,45 @@ else:
     st.write("#### :orange[Social Security]")
     col1, col2, col3 = st.columns(3, gap="large", vertical_alignment="top")
     with col1:
+        dob0 = kz.getCaseKey("dob0")
+        specialcase0 = dob0.endswith("01") or dob0.endswith("02")
         msg1 = "This is the **monthly** amount at Full Retirement Age (FRA)."
         msg2 = "Starting age of benefits in years and months."
         getIntInput(0, "ssAmt", "**monthly** PIA amount (in today's \\$)", helpmsg=msg1)
-        st.markdown(f"Use this [tool](https://ssa.tools/calculator) to get {kz.getCaseKey('iname0')}'s PIA.")
+        iname0 = kz.getCaseKey("iname0")
+        st.markdown(f"""Use this
+[tool](https://ssa.tools/calculator#integration=owlplanner.streamlit.app&dob={dob0}&useridx={iname0})
+to get {iname0}'s PIA.""")
         incol1, incol2 = st.columns(2, gap="large", vertical_alignment="top")
         with incol1:
             kz.initCaseKey("ssAge_m0", 0)
             maxyear = 70 if kz.getCaseKey("ssAge_m0") == 0 else 69
-            ret = getIntInput(0, "ssAge_y", "claiming at age...", 67, msg2, max_val=maxyear)
+            minyear = 62 if kz.getCaseKey("ssAge_m0") > 0 or specialcase0 else 63
+            ret = getIntInput(0, "ssAge_y", "claiming at age...", 67, msg2, min_val=minyear, max_val=maxyear)
         with incol2:
             maxmonth = 0 if ret == 70 else 11
-            getIntInput(0, "ssAge_m", "...and month(s)", 0, msg2, max_val=maxmonth, prompt=False)
+            minmonth = 1 if ret == 62 else 0
+            getIntInput(0, "ssAge_m", "...and month(s)", 0, msg2, min_val=minmonth, max_val=maxmonth, prompt=False)
 
     with col2:
         if kz.getCaseKey("status") == "married":
+            dob1 = kz.getCaseKey("dob1")
+            specialcase1 = dob1.endswith("01") or dob1.endswith("02")
             getIntInput(1, "ssAmt", "**monthly** PIA amount (in today's \\$)", helpmsg=msg1)
-            st.markdown(f"Use this [tool](https://ssa.tools/calculator) to get {kz.getCaseKey('iname1')}'s PIA.")
+            iname1 = kz.getCaseKey("iname1")
+            st.markdown(f"""Use this
+[tool](https://ssa.tools/calculator#integration=owlplanner.streamlit.app&dob={dob1}&useridx={iname1})
+to get {iname1}'s PIA.""")
             incol1, incol2 = st.columns(2, gap="large", vertical_alignment="top")
             with incol1:
                 kz.initCaseKey("ssAge_m1", 0)
                 maxyear = 70 if kz.getCaseKey("ssAge_m1") == 0 else 69
-                ret = getIntInput(1, "ssAge_y", "claiming at age...", 67, msg2, max_val=maxyear)
+                minyear = 62 if kz.getCaseKey("ssAge_m0") > 0 or specialcase1 else 63
+                ret = getIntInput(1, "ssAge_y", "claiming at age...", 67, msg2, min_val=minyear, max_val=maxyear)
             with incol2:
                 maxmonth = 0 if ret == 70 else 11
-                getIntInput(1, "ssAge_m", "...and month(s)", 0, msg2, max_val=maxmonth, prompt=False)
+                minmonth = 1 if ret == 62 else 0
+                getIntInput(1, "ssAge_m", "...and month(s)", 0, msg2, min_val=minmonth, max_val=maxmonth, prompt=False)
 
     st.divider()
     st.write("#### :orange[Pension]")